Simplify VerifyObjectBucketCleanup() by using early returns.
This just reduced the indentation level of the main code in
VerifyObjectBucketCleanup() by two levels, by using early returns.

It reduces indentation in VerifyObjectBucketCleanup() even further
by changing order, so that the last unindented block is just the good path.

This does not change the behavior but is just (imho) more readable.

View with "git show -w -U10" to see that it is hardly any diff really.

Signed-off-by: Michael Adam <obnox@redhat.com>