Don't require minikube for building.
`make cli` and `make gen-olm` run `openshift-sdk run ...` which requires a
kubeconfig. But these don't require a real kubernetes setup.
Up to now, this target relied on minikube being started.

This patch modifies the Makefile to use an empty kubeconfig file
for the invocations for building.

The `run` target still require a valid kubernetes environment and is not
modified.

Fixes: #325

Signed-off-by: Michael Adam <obnox@redhat.com>