Adding DB dump functionality
Adding a CLI command to dump the db.
The db dump is implemented for both mongodb and postgrse  db.
Implementation differs between the db types, as each is executing the
dump itself in a different manner, however the resulting resource is a
tar file of the db dump.
The syntax is "noobaa db-dump".
To get a db dump while using the diagnose command, add the flag "--db-dump"

Fixed: #169

Signed-off-by: Barak Sason Rofman <sason922@gmail.com>