Noobaa/Bucket Logging : syslog service for log transfer
Whenever an object operation comes to endpoint, handle_request
function process it and logs it to its /var/log/messages.

We need these logs to be sent on noobaa core so that we can process
it further and accordingly write it to log bucket as log objects.

In this code change we are creating a service noobaa-syslog. We will
send logs from end point to this service which in turn deliver it to
noobaa core.

Credit goes to aprinzse@redhat.com for the code contribution.

Signed-off-by: Ashish Pandey <aspandey@redhat.com>