added preStop hook to DB statefulset
The preStop hook will call pg_ctl with -m fast flag - this will rollback all active transactions and stop the server.
by default postgres will wait for all active connections to complete before exiting. the default grace period of kuberentes for pod shutdown is 30 seconds and after that it sends SIGKILL to the process. this can lead to DB failures

Signed-off-by: Danny Zaken <dannyzaken@gmail.com>