Azure kms handling for noobaa
This patch provides the support for azure keyvault.

We are using "libopenstorage/secrets" as the wrapper
package to integrate with different kms and package
provides the abstraction over several kms. It also
provides the integraton support for azure and helps
communication with azure key vault.

We are required to provide the definition for house
keeping calls registered calls with libopenstorage/secrets.

"libopenstorage/secrets" does the creation of client
handle based on the details provided in configmap.
The certificate details present in the secret are
preserved inside a temp file and used to establish
the connection with azure key vault as of now.

Below are the connection details that are going to be
populated on Noobaa CR by `noobaa_system_reconciler`
at ocs side and this is the ocs code where connectiondetails
on Noobaa CR are built: https://github.com/red-hat-storage/ocs-operator/blob/2d082fc4c1ac4cec961406053cece448f4b07684/controllers/storagecluster/noobaa_system_reconciler.go#L249

ex: configmap data:
```
data:
  AZURE_CERT_SECRET_NAME: azure-ocs-ffwc9o1j
  AZURE_CLIENT_ID: az-client-id1
  AZURE_TENANT_ID: az-tenant-id1
  AZURE_VAULT_URL: az-valut-url1
  KMS_PROVIDER: azure-kv
  KMS_SERVICE_NAME: kms-conn-azure1
```

Signed-off-by: Vinayakswami Hariharmath <vharihar@redhat.com>