Remove KMS condition status Init checks from tests
- Removing False positive verification of KMS condition status Init

The Init condition status is, in most cases, a state that is short in time, and the tests are missing the timing, hence failing.
We do not need this case as corev1.ConditionStatus = "Sync"
will never happen if Init did not, and we will fail there.

Signed-off-by: liranmauda <liran.mauda@gmail.com>
(cherry picked from commit de03454bb2142294da55b0bd5d047916deefd671)